Revert "Merge branch 'TransportSecure'"
This reverts commit a6872c50539a35d99bb9b6f33ddce8f585adbc41, reversing
changes made to ff20f682fc01835487246801e6f99fccd67bae8e.
