+++ common/net/minecraft/src/buildcraft/transport/PipeLogic.java
@@ -40,9 +40,6 @@ public class PipeLogic implements IDropControlInventory {
 		return false;
 	}
 
-	/// OWNERSHIP
-	public boolean isSecure() { return false; }
-	
 	public void writeToNBT(NBTTagCompound nbttagcompound) {
 
 	}
