--- common/net/minecraft/src/buildcraft/transport/PipeLogicSteel.java
@@ -1,5 +1,0 @@
-package net.minecraft.src.buildcraft.transport;
-
-public class PipeLogicSteel extends PipeLogic {
-	@Override public boolean isSecure() { return true; }
-}
