+++ common/net/minecraft/src/buildcraft/transport/PipeLogicWood.java
@@ -122,6 +122,4 @@ public class PipeLogicWood extends PipeLogic {
 		if (!APIProxy.isClient(worldObj))
 			switchSourceIfNeeded();
 	}
-	
-	@Override public boolean isSecure() { return true; }
 }
