+++ common/net/minecraft/src/buildcraft/transport/TileGenericPipe.java
@@ -34,7 +34,6 @@ import net.minecraft.src.buildcraft.api.PowerProvider;
 import net.minecraft.src.buildcraft.api.SafeTimeTracker;
 import net.minecraft.src.buildcraft.api.TileNetworkData;
 import net.minecraft.src.buildcraft.api.Trigger;
-import net.minecraft.src.buildcraft.api.pipes.IOwnable;
 import net.minecraft.src.buildcraft.core.BlockIndex;
 import net.minecraft.src.buildcraft.core.CoreProxy;
 import net.minecraft.src.buildcraft.core.IDropControlInventory;
@@ -51,8 +50,7 @@ import net.minecraft.src.buildcraft.core.network.PacketUpdate;
 
 public class TileGenericPipe extends TileEntity implements IPowerReceptor,
 		ILiquidContainer, ISpecialInventory, IPipeEntry, ISynchronizedTile,
-		IOverrideDefaultTriggers, ITileBufferHolder, IPipeConnection, IDropControlInventory,
-		IOwnable {
+		IOverrideDefaultTriggers, ITileBufferHolder, IPipeConnection, IDropControlInventory {
 
 	public TileBuffer [] tileBuffer;
 	public boolean [] pipeConnectionsBuffer = new boolean [6];
@@ -65,14 +63,14 @@ public class TileGenericPipe extends TileEntity implements IPowerReceptor,
 
 	@TileNetworkData public int pipeId = -1;
 
-	public TileGenericPipe () {}
+	public TileGenericPipe () {
+
+	}
 
 	@Override
 	public void writeToNBT(NBTTagCompound nbttagcompound) {
 		super.writeToNBT(nbttagcompound);
 
-		if(owner != null && !owner.isEmpty())
-			nbttagcompound.setString("owner", owner);
 		if (pipe != null) {
 			nbttagcompound.setInteger("pipeId", pipe.itemID);
 			pipe.writeToNBT(nbttagcompound);
@@ -83,8 +81,6 @@ public class TileGenericPipe extends TileEntity implements IPowerReceptor,
 	public void readFromNBT(NBTTagCompound nbttagcompound) {
 		super.readFromNBT(nbttagcompound);
 
-		if(nbttagcompound.hasKey("owner"))
-			owner = nbttagcompound.getString("owner");
 		int key = nbttagcompound.getInteger("pipeId");
 		if (key > 0) {
 			pipe = BlockGenericPipe.createPipe(key);
@@ -122,26 +118,6 @@ public class TileGenericPipe extends TileEntity implements IPowerReceptor,
 
 	public boolean initialized = false;
 
-	/// OWNERSHIP
-	private String owner;
-	
-	@Override
-	public boolean isSecure() {
-		if(pipe != null)
-			return pipe.isSecure();
-		else
-			return false;
-	}
-	@Override
-	public String getOwnerName() {
-		return owner;
-	}
-	@Override
-	public void setOwner(EntityPlayer player) {
-		owner = player.username;
-	}
-	
-	/// UPDATING
 	@Override
 	public void updateEntity () {
 		if (!initialized) {
